move userPreferences to methods folder
