+++ app/lib/rocketchat/methods/logout.ts
@@ -10,7 +10,7 @@ import RocketChat from '..';
 import { useSsl } from '../../../utils/url';
 import log from '../../../utils/log';
 import { E2E_PRIVATE_KEY, E2E_PUBLIC_KEY, E2E_RANDOM_PASSWORD_KEY } from '../../constants';
-import UserPreferences from '../../userPreferences';
+import UserPreferences from '../../methods/userPreferences';
 import { ICertificate, IRocketChat } from '../../../definitions';
 import sdk from '../services/sdk';
