+++ app/sagas/encryption.js
@@ -8,7 +8,7 @@ import Navigation from '../lib/Navigation';
 import { E2E_BANNER_TYPE, E2E_PRIVATE_KEY, E2E_PUBLIC_KEY, E2E_RANDOM_PASSWORD_KEY } from '../lib/constants';
 import database from '../lib/database';
 import RocketChat from '../lib/rocketchat';
-import UserPreferences from '../lib/userPreferences';
+import UserPreferences from '../lib/methods/userPreferences';
 import { getUserSelector } from '../selectors/login';
 import { showErrorAlert } from '../utils/info';
 import I18n from '../i18n';
