+++ app/utils/theme.ts
@@ -5,10 +5,9 @@ import setRootViewColor from 'rn-root-view';
 import { IThemePreference, TThemeMode } from '../definitions/ITheme';
 import { themes } from '../constants/colors';
 import { isAndroid } from './deviceInfo';
-import UserPreferences from '../lib/userPreferences';
+import UserPreferences from '../lib/methods/userPreferences';
 import { THEME_PREFERENCES_KEY } from '../lib/constants';
 import { TSupportedThemes } from '../theme';
-
 
 let themeListener: { remove: () => void } | null;
