+++ app/views/NewServerView/index.tsx
@@ -20,7 +20,7 @@ import I18n from '../../i18n';
 import database from '../../lib/database';
 import { sanitizeLikeString } from '../../lib/database/utils';
 import RocketChat from '../../lib/rocketchat';
-import UserPreferences from '../../lib/userPreferences';
+import UserPreferences from '../../lib/methods/userPreferences';
 import { OutsideParamList } from '../../stacks/types';
 import { withTheme } from '../../theme';
 import { isTablet } from '../../utils/deviceInfo';
